pyproject.toml
[tool.poetry]
name = "netsurf_rss"
version = "1.0.0"
description = "Giving an awesome browser's news feed the RSS feed it always deserved."
authors = ["Kyle Williams <kyle.anthony.williams2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
rfeed = "^1.1.1"
Flask = "^2.0.1"
selectolax = "^0.3.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
